Add support for Jinja tempating in bq_queries
With Jinja templating we can specify Jinja syntax in bq_queires.
One particularly useful element is for loops (proposed implementation).

We use new template.* section of parameters to specify something like
`template.cohort_days=1,2,3,5,7` and it will be automatically converted to
array of values [1, 2, 3, 5, 7]. Based on this values query will be
expanded (with help of new `_expand_jinja` method) and will include as
many rows as we have in the resulting array.

Change-Id: Ide9c05cde79553e4156be576048d121be269f149