Handle case of single element for expanding jinja
Jijna template in BigQueryExecutor now only supports *for* loops. When
we want to expand only a single parameter we either face an error (`int`
is not iterable`) or undesired behaviour (iterating over `str` iterates
over each character in the string).
To solve this we need to convert a single element to a list in
`_expand_jinja` method.

Change-Id: I86b7571b21a13bb1b0e8dad96212ca7af72a23b2