Add initializing parameters in cli.gaarf
Current implementation of `ParamsParser` initializes runtime parameters
(such as `:YYYYMMDD`) when calling `build` method from
`GaarfConfigBuilder`. This results in the fact that saving config with
`ConfigSaver` always returns fixed values for `:YYYYMDDD` parameters
being converted to actual dates (i.e. 2022-01-01) which makes it harder
to re-use this config.

In order to solve this do the following:

* Remove call to `convert_date` from `_identify_param_pair` method of
  `ParamsParser`- all the params taken from CLI arguments remain
  un-initialized.
* Add `initialize_runtime_parameters` function that operates on
  `GaaafConfig` and `GaarfBqConfig` and initilize any runtime parameters
  it can find the these configs
* Call `initialize_runtime_parameters` in `cli.gaarf` right after config
  is saved to yaml file. This ways yaml file contains all parameters
  that are un-initialized.

Change-Id: Ie43b6312dbdd2667509b47624c822ee1fd6984d9