Add support for SqlAlchemy writer
SqlAlchemyWriter allows to re-use various DB engines provided by
`sqlalchemy` (i.e. Postgres).

In order to use it we need to provide `--output=sqldb` to `gaarf`
and `--sqldb.connection_string` a connection string used to
initialize connection to a desired DB.

Example usage:

```
gaarf path/to/query.sql --account=... --output=sqldb
--sqldb.connection_string="postgresql://user:password@host:port/dbname"
```

While providing a raw connection_string is a suboptimal solution, it's
fast and it's working now - we will definitely refactor it in the
future.

Change-Id: I77cc28b52fd1952775e973dc962f3391ba69bf11