[gaarf-py] Convert non-exiting attributes of GoogleAdsRow to None
One particular segments `segments.sk_ad_network_conversion_value` is
treated as 0 while not being actually present in the `GoogleAdsRow`
object. There's a huge difference between 0 and NULL values for this
particular segment; we apply `HasField` check on the field and replace
default 0 value with None if the field is not present.

Change-Id: Ib604f1fccd512664cf74946f45f9cfb8404283dc