[gaarf-py] Format macros inside virtual columns
If macro is use inside a virtual column it will be converted to passed
macro value (previously it will be treated literally, i.e. if query
contains virtual column `{value}` it will end up as `{value}` in the
output.
The logic of applying macros has changed - macro expansion is perform
during the virtual column definition.

Change-Id: I381d6c248ae23fe766a1ceebaf07c9149dc8df51