[gaarf-py] Ensure that ocid is always converted to scalar
Current behaviour of built-in query `ocid_mapping` return a list with a
single element for `ocid` column; but due to writing behavior of
gaarf-py it's always converted to string.
By getting only the first element of found `ocid` we can guarantee the
same behaviour both for CLI and library use cases.

Change-Id: I36c2bb07fffc2641807378e35cc945b614d0621f